requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pandas>=2.0
numpy>=1.24
yfinance>=0.2
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
tensorflow>=2.15
keras>=3.0

pytest>=8.0
httpx>=0.26

langchain>=0.2
langchain-community>=0.2
langchain-chroma>=0.1
chromadb>=0.5
sentence-transformers>=3.0
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.3
tabulate>=0.9
